Change renderExplorerUrl to utils
